Add files via upload
The following Python code documents the research and testing conducted to identify the correct invariance and embedding of physics and the methods used to segment the reduced MNIST data. This is being uploaded solely as documentation of the work leading to the final paper, and none of the code was used for the outcome.